- add chat screen (ChatList & ChatPane components) - define and add routes to client (install react-router-dom) - add dynamic routing to cater for screen sizes in '/chat' route (install react-media) - implement route work flow from Home screen (signup/login component(s)) to chat screen display - implement route work flow to stop unauthorized access of Chat screen
