+++ package.json
@@ -26,6 +26,7 @@
     "mongodb": "^3.5.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-media": "^1.10.0",
     "react-router-dom": "^5.1.2",
     "redis": "^2.8.0",
     "rxjs": "^6.5.4",
@@ -43,6 +44,7 @@
     "@types/node": "^13.7.0",
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
+    "@types/react-router-dom": "^5.1.3",
     "@types/redis": "^2.8.14",
     "@types/socket.io": "^2.1.4",
     "@types/socket.io-client": "^1.4.32",
